PRSS refactor
