+++ Cargo.toml
@@ -9,8 +9,8 @@ default = ["debug", "cli"]
 cli = ["comfy-table", "clap", "enable-serde", "metrics-tracing-context", "metrics-util", "tracing-subscriber", "web-app"]
 debug = []
 enable-serde = ["serde", "serde_json"]
-protocol-perf-testing = ["turn-off-prss"]
-turn-off-prss = []
+protocol-perf-testing = ["no-prss"]
+no-prss = []
 web-app = ["tokio", "tokio-stream", "tokio-util", "axum", "axum-server", "hyper", "hyper-tls", "tower", "tower-http"]
 self-signed-certs = ["hyper-tls"]
 test-fixture = []
