+++ src/protocol/context/mod.rs
@@ -81,7 +81,10 @@ mod tests {
     use crate::ff::{Field, Fp31};
     use crate::helpers::Direction;
     use crate::protocol::malicious::Step::MaliciousProtocol;
-    use crate::protocol::prss::SharedRandomness;
+    #[cfg(feature = "no-prss")]
+    use crate::protocol::no_prss::SharedRandomness;
+    #[cfg(not(feature = "no-prss"))]
+    use crate::protocol::use_prss::SharedRandomness;
     use crate::protocol::RecordId;
     use crate::secret_sharing::{MaliciousReplicated, Replicated};
     use crate::telemetry::metrics::{
